Add an initial version of ucored(8)
No documentation yet, and it isn't yet fully functional.  This version will
accept devd notifications and pipe all of the relevant metadata over to
ucored(8) to act on it, but the policy bits aren't yet hooked up.

The next step will introduce some lua that reads our config and handles
each entry that comes in.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>