ucored: initial implementation of the lua-side
This actually has a number of decent features; the default behavior is to
move all coredumps over to /var/ucrash and add the pid to the filename to
de-conflict core filenames.  At the current time, we support the following
symbols in 'move' destination paths:

 - %j = jid
 - %n = original filename
 - %P = parent pid
 - %p = pid
 - %s = signal

The move is executed by a rename(2) where possible, copy_file_range(2) where
it is not.  One can also configure discard/ignore actions that are
self-explanatory.  ucored grew another option that I hadn't considered
previously to execute a third party script to handle it.  That one should
have the same interface as ucored.lua: it returns a handler function that
will take a ucore.  It has the full lua library available to it, and can
additionally include our 'core' module.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>