ucored: allow us to be launched via, e.g., inetd
I would think the startup costs of reading the lua script and configuration
aren't necessarily worth it, but maybe someone would want it.  Note that the
kern.coredump_devctl sysctl needs to be enabled manually; normally this is
done by ucored(8) before it heads into its event loop, but we obviously
don't get the chance to enable it if we're relying on ucore-shuttle -> inetd
to spawn us off.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>