libucore: re-align ucore_data to max_align_t
As of today we only send strings across, but I don't want to assume we won't
decide to add some more complex structure.  Increase the alignment
requirements to be more accommodating to arbitrary data.

While we're here, ensure we zero out any padding we may have between the
ucore_data_hdr and the subsequent payload to be safe.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>