ucored, ucore-shuttle: check credentials of the other side
From ucored's perspective, we don't want to receive "coredump" notifications
from non-root because we'll often be moving files described in the note to
a common location.  We could perhaps relax this to checking later that it's
either root *or* the uid of the file owner, but for now we only want to
accept notifications from devd anyways.  The admin needs to be careful about
exposing the ucored socket to jails.

From ucore-shuttle's perspective, it wants to avoid exposing core details
to unprivileged users just in case system policy would otherwise prevent
them from learning details.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>