ucored: actually check for EOF
Right now, a client can connect and hangup and we'd just simply never
discard their state.  ucore_fetch() would never actually see the socket as
having enough data (because kevent() just kicked it back to observe EV_EOF),
so we'd enter into a vicious cycle where we're just continuously rearming
our EVFILT_READ event.  This is an unfortunate side-effect of the protocol
defining exactly how many bytes we should expect to see next, we won't
really attempt to read() if we think it'd be premature.

Manage this scenario better by actually checking for EV_EOF.  Even for our
normal scenarios, if EV_EOF was set then we should avoid sending the ack
that we just setup because we know it'll error out.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>