ucored: close all sockets on fork()
We don't currently, but in the future we'll fork() to handle cores that we
have received either from a shuttle or from the kernel directly to avoid
jamming up processing of other incoming cores.  Given that processing may
take a while, particularly if we end up going over the network, let's be
proactive and mark all of our sockets for CLOFORK to avoid holding anything
open.

In the future, when we actually start forking, we'll probably want to
closefrom(2) around if SOCK_CLOFORK == 0.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>