libucore, ucored: rework our kqueue to use the notion of a readable
Refactor to avoid spreading knowledge of the kqueue itself out too far.
We'll use this later to add a readable for the /dev/ucore reader, to
avoid embedding knowledge of how we handle those into ucored.

The lifetime of a ucore_readable is managed by the readable object, not
by the kqueue processor.  For convenience, we'll just work with embedded
struct ucore_readable rather than tracking context pointers.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>