libucore: add support for /dev/ucore
The API here is rather simple; one opens the device with
libucore_dev_open(), which is passed a handler function that it will
invoke when a new ucore is grabbed from ucoredev.  Then,
libucore_dev_readable() returns a `ucore_readable *` that contains
the details necessary to poll() the socket or enroll it in a kqueue, and
we call its r_read() callback when we think it should have data
available.

If r_read() is passed (size_t)-1 for the avail parameter, it'll use
FIONREAD to fetch the availability to avoid blocking before we run out
of new fds.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>