kmod: frame out our coredumps with additional information
We're using the same binary format here as the ucore-shuttle for
convenience.  Future work may support forwarding on this exact format
to a remote ucored(8) from our local ucored(8), so avoiding deviation
is useful.

For now we're only adding the ucore header and jail information, but we
also need to grab the command name from the text vnode and perhaps also
the process' working directory.  Perhaps we'll find other context useful
to grab as well down the road.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>