ucored: add support for /dev/ucore
This switches wholesale to just using ucoredev if it's available, rather
than keeping a socket open for ucore-shuttle to try and send us
anything.  With the ucore_readable design, all we have to do is add it
to the kqueue and the main loop takes care of the rest.

Some other parts are adjusted to account for the differences between
shmfd dumps and vnode dumps.  Notably, we don't actually have a backing
file, so we may have to make up a name for %n and we need a
copy_file_range() backup, since that syscall only works on regular
files.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>