kmod, ucore-shuttle: add more details to our header
In a traditional coredump setup, we have uid/gid information along with
other details that we need to create the dump.  We don't get any of
those with the shmfd, so we need to ensure our frame includes them.

The other end of our socket (ucore-shuttle) is assumed to be privileged,
so we'll use these in the future to create our new core since we won't
always have stat(2) details.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>