kmod: clean up some globals and device open/close
Organize some of this into a softc, just to be a little more tidy.
Start tracking if the device is open and don't allow it to be opened
more than once, and start blocking unload if the device is still open.

Now that we have a better handle of device open/close, manage the
select/kqueue state around then rather than at module load/unload time.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>