ucored: pull our configuration from /usr/local/etc/ucored.conf
For the time being, we only support "filters" at the top-level.  Each
filter in the array of filters has two keys: match and execute.  match
has the attributes we want to filter on mapped to regular expressions
that we want to match against, and execute describes one or more actions
to apply.  For instance, one could write a filter that executes a move
off into /var/ucrash/$jail, then executes some script or pipes the core
into another command.

With this, we install a default configuration that simply moves all
cores off into /var/ucrash and appends the pid to the filename to
uniquify it.  For future work, it would be good to match the feature
set of core(5), which mainly means we're missing:
 - %H -> hostname
 - %I -> index
 - %u -> uid

%N behaves differently in ucored(8), which should be fixed:  we use it
as the filename, but core(5) defines it as the process name.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>