ucored: lua: unbreak the pipe action now that ucored(8) forks
When ucored(8) started to fork, we registered a SIGCHLD handler and
moved on with life.  That's incompatible with how we manage our pipe
action, though, which wants to wait on the status before moving on with
potentially other actions.

In the unlikely event that we're running a handler in ucored(8) itself
(because fork(2) failed), we'll signal the main loop to check for any
children that we might have missed a SIGCHLD for.  At the sane time,
let's fix the problem with async-signal-unsafe stuff in our SIGCHLD
handler by signalling there as well.  We should return to kqueue()
pretty quickly nowadays, as our handlers will typically fork off.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>