ucored: carefully blow away core data when we're done with it
The metadata is a bit less sensitive, but we definitely want to clear
out buffers that might've had the core itself.  We try to separate
handling of cores with fork(), but we'll still allow a core to be
processed in the main ucored(8) process if we encountered a fork()
failure for any reason, just to avoid losing a core.

Blow away the buffers and any associated metadata that we might've
collected from the core, to be safe.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>