libucore: add a notion of data-only open, use it in ucoredev(4) open
Our traditional ucored_client only opens data segments, so it's fine and
just ignores the new hint.

ucoredev(4) cores, on the other hand, will use it to copy-out the core
data to a new memfd.  In theory, this will be used to pipe the core off
somewhere else to a command that may or may not find it weird to be
sensitive to the file offset when we executed it

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>