ucored: add a timer to periodically close connections
To avoid the possibility of something going terribly awry on the other
side and just wasting our resources, we'll do a timeout every 60 seconds
or so.  We'll only keep the timer active when we have active clients,
since we don't really need a periodic wakeup for the vast majority of
the time when nothing is generated a coredump.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>