ucoredev: fix our init_fn definition
The final version of this that landed in -CURRENT no longer passes the
compression format through separately.  The coredump_params has the
compressor handle in it, and we now have an API to grab the format from
the handle.

While we're here, drop a debugging statement that shouldn't have made
it in.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>