ucored: address some clang-analyze complaints
Some of these are just cosmetic, but there's a few functional gems
hiding in there -- a double-free, a case of using errno instead of
serrno, and a few potential memory leaks in failure cases.

This also moves the ucored_state enum to ucored_client.c, because that's
actually just client state that never gets exposed elsewhere.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>