kmod: downgrade the ucom_data_hdr size down to a uint32_t
Our current MAXSEGSZ is 4k, and I don't really see us needing to copy
more than that in a single data segment, particularly in the kernel.  If
we find a need to bump it, I have a hard time believing we'd find 4GB
worth of piece of data to export that didn't make more sense being in
the core itself somehow.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>