ucore-shuttle: note the packing difference for data segments
ucore-shuttle, since it's operating over a socket, will expect the
receiving side to receive into an arbitrary buffer and copy it over into
a struct ucore_data -- thus, it doesn't pack the padding into the socket
between header/data.   ucored(8) does account for this, so it's not a
problem.  I suspect nobody else would write a receiver for ucore-shuttle
data.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>