build: statically link ucored programs
Given that we take over coredump handling for the system on FreeBSD 15.0
and up, let's avoid the potential for system funkery.  In particular, we
don't want to end up in a state where a dependency can break us, e.g.,
in the middle of or as the result of an upgrade that didn't end up
rebuilding ucored.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>